backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
celery==5.5.3
redis==6.4.0
requests==2.32.3
django-corsheaders==4.4.0
google-generativeai==0.8.3
strawberry-graphql-django==0.46.0
strawberry-graphql==0.239.0